APPDEV-4126 : Extole Cron job fixes
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@38603 842219ff-9341-0410-a943-573460bdd4de

+++ projects/ERPServices/src/com/freshdirect/referral/extole/model/FDRafCreditModel.java
@@ -54,7 +54,7 @@ public class FDRafCreditModel implements Serializable{
 	
 	private String rewardSetId;
 		
-	private int rewardValue;
+	private double rewardValue;
 		
 	private String rewardDetail;
 	
@@ -287,11 +287,11 @@ public class FDRafCreditModel implements Serializable{
 		this.rewardSetName = rewardSetName;
 	}
 
-	public int getRewardValue() {
+	public double getRewardValue() {
 		return rewardValue;
 	}
 
-	public void setRewardValue(int rewardValue) {
+	public void setRewardValue(double rewardValue) {
 		this.rewardValue = rewardValue;
 	}
